requirements.txt
# Core dependencies
RPi.GPIO>=0.7.1
spidev>=3.5
smbus2>=0.4.2
pyserial>=3.5

# RFID/NFC
mfrc522>=0.0.7
pn532>=1.1
nfcpy>=1.0.4

# Radio & RF
RFM69>=0.4.1

# Display
Adafruit-SSD1306>=1.6.2
Adafruit-GPIO>=1.0.3
Pillow>=9.0.0

# IR
python-lirc>=0.0.1

# USB/HID
pyusb>=1.2.1
hid>=1.0.5

# Network & Security
scapy>=2.4.5
bluepy>=1.3.0
pybluez>=0.23

# UI & CLI
Click>=8.0.0
rich>=10.0.0
prompt-toolkit>=3.0.0

# Utilities
python-dotenv>=0.19.0
pyyaml>=6.0
requests>=2.27.0
numpy>=1.21.0
